chore: bump rust toolchain to 1.67 (#997)

+++ core/executor/benches/mock.rs
@@ -154,5 +154,5 @@ fn time_now() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u64
+        .as_secs()
 }
